+++ src/equations/ideal_glm_mhd_multicomponent_1d.jl
@@ -134,7 +134,7 @@ end
 
 
 # Calculate 1D flux in for a single point
-@inline function flux(u, orientation, equations::IdealGlmMhdMulticomponentEquations1D)
+@inline function flux(u, orientation::Integer, equations::IdealGlmMhdMulticomponentEquations1D)
   rho_v1, rho_v2, rho_v3, rho_e, B1, B2, B3 = u
 
   rho = density(u, equations)
